refactor: move bcrypt-adapter files to a folder
